requirements.txt
line-bot-sdk
django
gunicorn
django-heroku
datetime
Pillow
django-import-export
apiai
pymysql
sqlalchemy
pandas
bs4
matplotlib